+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Trait based type parameter for system reset, pmu and base extension
+- Feature `integer-impls` to allow fast prototyping with sbi-rt crate
 - Feature `legacy` to gate SBI legacy extension
 - Documents on various functions